refactor(bencode): rename bencoding module to bencode
