+++ bitcrawler-proto/src/krpc/query.rs
@@ -1,7 +1,7 @@
 use std::collections::HashMap;
 
 use crate::{
-    bencoding::{BencodeDict, BencodeValue},
+    bencode::{BencodeDict, BencodeValue},
     kademlia::NodeId,
 };
