+++ bitcrawler-proto/src/lib.rs
@@ -1,3 +1,3 @@
-pub mod bencoding;
+pub mod bencode;
 pub mod kademlia;
 //pub mod krpc;
